+++ projects/pastanaga-angular/src/lib/table/table-sortable-header/table-sortable-header.component.spec.ts
@@ -145,7 +145,6 @@ describe('TableSortableHeaderComponent', () => {
             component.ngAfterViewInit();
             expect(component.sortMenuPosition).toEqual({
                 position: 'absolute',
-                left: 10,
                 top: 130,
             });
         });