+++ CHANGELOG.md
@@ -5,6 +5,9 @@
     - Translations are now added from any module using the `addTranslations` static method
     - `mergeTranslations` is not available anymore
     - `PA_TRANSLATIONS` cannot be provided anymore
+    
+### BugFix
+- **Sortable Table Header**: Prevent sortable table header mobile dropdown to impact table rows [mpellerin42]
 
 # 2.27.1 (2021-04-21)