+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### BugFix
 - **Sortable Table Header**: Prevent sortable table header mobile dropdown to impact table rows [mpellerin42]
 - **Expander**: Update content height on expand to prevent overlap when window is resized while collapsed [mpellerin42]
+- **Typography**: Update display line-height to prevent low and high characters from being cut [mpellerin42]
 
 # 2.27.1 (2021-04-21)